+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -94,7 +94,7 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-02T09:50:17.688Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-02T13:58:47.485Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")