+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT3VVXSBZTMBZXRG971VY7PR"
+        "01KT4A32HKSFA74G3R5JTDKTWK"
     )
     r"""The unique identifier of the data source"""