+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT3VVXSDHPSP4P7DAVBNNV8X"
+        "01KT4A32HPKBMBD3FSFND69C92"
     )
     r"""The unique identifier of the data source"""