+++ packages/orq-rc/RELEASES.md
@@ -8389,3 +8389,13 @@ Based on:
 - [python v4.10.0-rc.69] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.69] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.69 - packages/orq-rc
+
+## 2026-06-02 14:01:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.769.2 (2.892.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.72] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.72] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.72 - packages/orq-rc